Refactor Function decorators for backend and frontend.
