+++ ivy_tests/test_ivy/helpers/decorators/backend/function_decorator_backend.py
@@ -1,5 +1,3 @@
-import inspect
-
 from hypothesis import strategies as st
 from typing import Callable, Any
 from ivy_tests.test_ivy.helpers.globals import TestData
@@ -55,10 +53,9 @@ class BackendFunctionHandler(FunctionHandler):
             container_flags=container_flags,
         )
 
-        self._init_possible_args()
-
-    def _init_possible_args(self):
-        self.possible_args = {
+    @property
+    def possible_args(self):
+        return {
             "ground_truth_backend": st.just(self.ground_truth_backend),
             "fn_name": st.just(self.test_data.fn_name),
             "test_data": self.test_data,
@@ -74,27 +71,8 @@ class BackendFunctionHandler(FunctionHandler):
             supported_device_dtypes=supported_device_dtypes,
         )
 
-    def _add_test_attrs_to_fn(self, fn: Callable[..., Any]):
+    def _add_test_attributes_to_test_function(self, fn: Callable[..., Any]):
         fn._is_ivy_backend_test = True
         fn.ground_truth_backend = self.ground_truth_backend
         fn.test_data = self.test_data
         return fn
-
-    def _update_given_kwargs(self, fn):
-        param_names = inspect.signature(fn).parameters.keys()
-
-        # Check if these arguments are being asked for
-        filtered_args = set(param_names).intersection(self.possible_args.keys())
-        for key in filtered_args:
-            self._given_kwargs[key] = self.possible_args[key]
-
-    def __call__(self, fn: Callable[..., Any]):
-        if self.is_hypothesis_test:
-            self._update_given_kwargs(fn)
-            wrapped_fn = self._wrap_with_hypothesis(fn)
-        else:
-            wrapped_fn = fn
-
-        self._add_test_attrs_to_fn(wrapped_fn)
-        self._handle_not_implemented(wrapped_fn)
-        return wrapped_fn
